Ways to Earn Money Online
There are different routes through which one can earn money online in India, however, the most preferred ways, where a majority of people are engaged, are listed below:

Become a Freelancer

The best and most reliable option is to become a freelancer. If you are a good programmer, designer, marketer, or a writer then you can find a lot of paid jobs. You need to have two skills to become a good freelancer. One is your core skill, and the second is marketing. You have to have brilliant communication skills to get clients. The Internet has significantly expanded opportunities to earn money working as a freelancer.

Become a Consultant

If you think you are capable enough and filled with knowledge that you can sell your advice and knowledge to people then it is something worth trying for. You don’t have to be super skilled in one field to become a consultant or teacher, you just have to be better than your student or client. Anyone with a core competitive skill can become a consultant and find clients online. 

Virtual Assistant

A virtual assistant is a personal assistant who can earn money by working online for someone without being physically present. The tasks you are expected to do is take care of websites, counseling, writing & proofreading, publishing content, marketing, coding, website & app development, research etc. It is quite interesting, one who is capable of doing it should give it a shot.

Become a Blogger

Becoming a blogger is a legitimate way to make money. Blogging is something that requires patience, diligence and discipline. It might take long, one or even two years building your blog, your brand, and your authority, before making any serious amount of money but at the end, it’s all worth the patience.

YouTuber

According to recent stats, we now watch more videos on YouTube than searches on Google. You will receive a percentage of the advertising revenue collected per 1,000 views. You can make a lot of money, and there are plenty of stories every week of more and more YouTubers making it their career.

Poker

Online poker is something everyone should try. Once the game is understood well, it will give a limitless fund flow. Online Poker takes some intensive study, you need to learn all of the tricks, plus rack skill, the endgame, the psychology of it, statistics, etc. And then you need to play other hard-core players and beat them. The best part of learning poker is that it will not only give the way to finance, but also give you some of the best life lessons. The average poker player, depending on stakes, can make as little as $10,000 a year up to $1,000,000 or more per year.